rft: rename Notification -> ExternalNotification
