+++ src/MaaWpfGui/Services/ExternalNotification/BarkNotificationProvider.cs
@@ -20,7 +20,7 @@ using MaaWpfGui.Models.ExternalNotification;
 using MaaWpfGui.Services.Web;
 using Serilog;
 
-namespace MaaWpfGui.Services.Notification;
+namespace MaaWpfGui.Services.ExternalNotification;
 
 public class BarkNotificationProvider(IHttpService httpService, BarkConfig bark) : IExternalNotificationProvider
 {
