+++ src/MaaWpfGui/Services/ExternalNotification/DiscordNotificationProvider.cs
@@ -21,7 +21,7 @@ using MaaWpfGui.Models.ExternalNotification;
 using MaaWpfGui.Services.Web;
 using Serilog;
 
-namespace MaaWpfGui.Services.Notification;
+namespace MaaWpfGui.Services.ExternalNotification;
 
 public class DiscordNotificationProvider(IHttpService httpService, DiscordConfig discord) : IExternalNotificationProvider
 {
