+++ src/MaaWpfGui/Services/ExternalNotification/GotifyNotificationProvider.cs
@@ -22,7 +22,7 @@ using MaaWpfGui.Models.ExternalNotification;
 using MaaWpfGui.Services.Web;
 using Serilog;
 
-namespace MaaWpfGui.Services.Notification;
+namespace MaaWpfGui.Services.ExternalNotification;
 
 /// <inheritdoc />
 public partial class GotifyNotificationProvider(IHttpService httpService, GotifyConfig gotify) : IExternalNotificationProvider
