+++ src/MaaWpfGui/Services/ExternalNotification/TelegramNotificationProvider.cs
@@ -24,7 +24,7 @@ using MaaWpfGui.Services.Web;
 using MaaWpfGui.ViewModels.UI;
 using Serilog;
 
-namespace MaaWpfGui.Services.Notification;
+namespace MaaWpfGui.Services.ExternalNotification;
 
 public class TelegramNotificationProvider(IHttpService httpService, TelegramConfig telegram) : IExternalNotificationProvider
 {
